requirements.txt
# requirements.txt
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pillow>=10.0.0
pytesseract>=0.3.0
groq>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
speechrecognition>=3.10.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
flask>=2.0.0
flask-cors>=3.0.0

flask-jwt-extended>=4.5.2
pymongo>=4.5.0
google-auth>=2.23.0
pyjwt>=2.8.0
flask-cors>=4.0.0
bcrypt>=4.0.1

# # Core Flask
# Flask==2.3.3
# flask-cors==4.0.0
# python-dotenv==1.0.0

# # Authentication
# flask-jwt-extended==4.5.2
# PyJWT==2.8.0
# google-auth==2.23.0
# google-auth-oauthlib==1.1.0
# bcrypt==4.0.1

# # Database
# pymongo==4.5.0
# dnspython==2.4.2

# # AI & Data Processing
# openai==1.3.0
# chromadb==0.4.15
# sentence-transformers==2.2.2
# PyPDF2==3.0.1
# python-docx==1.1.0
# Pillow==10.1.0
# pytesseract==0.3.10
# groq==0.3.0
# numpy==1.24.3
# pandas==2.1.3

# # Utilities
# requests==2.31.0
# Werkzeug==2.3.7
# Jinja2==3.1.2
# click==8.1.7

